+++ backend/webtoon-core/src/main/java/com/webtoon/core/comment/service/CommentLikeDislikeService.java
@@ -2,14 +2,14 @@ package com.webtoon.core.comment.service;
 
 import com.webtoon.core.comment.domain.Comment;
 import com.webtoon.core.comment.domain.CommentDislike;
-import com.webtoon.core.comment.domain.CommentDislikeRepository;
+import com.webtoon.core.comment.repository.CommentDislikeRepository;
 import com.webtoon.core.comment.domain.CommentLike;
-import com.webtoon.core.comment.domain.CommentLikeRepository;
-import com.webtoon.core.comment.domain.CommentRepository;
+import com.webtoon.core.comment.repository.CommentLikeRepository;
+import com.webtoon.core.comment.repository.CommentRepository;
 import com.webtoon.core.comment.dto.CommentLikeDislikeCountResponse;
 import com.webtoon.core.common.exception.ApplicationException;
 import com.webtoon.core.user.domain.User;
-import com.webtoon.core.user.domain.UserRepository;
+import com.webtoon.core.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
