+++ backend/webtoon-core/src/main/java/com/webtoon/core/episode/repository/EpisodeRepository.java
@@ -1,5 +1,6 @@
-package com.webtoon.core.episode.domain;
+package com.webtoon.core.episode.repository;
 
+import com.webtoon.core.episode.domain.Episode;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
