+++ backend/webtoon-core/src/main/java/com/webtoon/core/webtoon/service/WebtoonService.java
@@ -8,9 +8,9 @@ import com.webtoon.core.webtoon.dto.WebtoonCreateRequest;
 import com.webtoon.core.webtoon.dto.WebtoonEditRequest;
 import com.webtoon.core.webtoon.dto.WebtoonMainPageResponse;
 import com.webtoon.core.user.domain.User;
-import com.webtoon.core.user.domain.UserRepository;
+import com.webtoon.core.user.repository.UserRepository;
 import com.webtoon.core.webtoon.domain.Webtoon;
-import com.webtoon.core.webtoon.domain.WebtoonRepository;
+import com.webtoon.core.webtoon.repository.WebtoonRepository;
 import com.webtoon.core.webtoon.dto.WebtoonResponse;
 import com.webtoon.core.webtoon.dto.WebtoonsMainPageResponse;
 import org.springframework.data.domain.Page;
