+++ backend/webtoon-core/src/test/java/com/webtoon/core/comment/service/CommentLikeDislikeServiceTest.java
@@ -2,13 +2,13 @@ package com.webtoon.core.comment.service;
 
 import com.webtoon.core.user.domain.User;
 import com.webtoon.core.user.domain.enums.Gender;
-import com.webtoon.core.user.domain.UserRepository;
+import com.webtoon.core.user.repository.UserRepository;
 import com.webtoon.core.comment.domain.Comment;
 import com.webtoon.core.comment.domain.CommentDislike;
 import com.webtoon.core.comment.domain.CommentLike;
-import com.webtoon.core.comment.domain.CommentDislikeRepository;
-import com.webtoon.core.comment.domain.CommentLikeRepository;
-import com.webtoon.core.comment.domain.CommentRepository;
+import com.webtoon.core.comment.repository.CommentDislikeRepository;
+import com.webtoon.core.comment.repository.CommentLikeRepository;
+import com.webtoon.core.comment.repository.CommentRepository;
 import com.webtoon.core.comment.dto.CommentLikeDislikeCountResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
