+++ backend/webtoon-core/src/test/java/com/webtoon/core/comment/service/CommentServiceTest.java
@@ -3,17 +3,17 @@ package com.webtoon.core.comment.service;
 import com.webtoon.core.comment.dto.CommentResponse;
 import com.webtoon.core.user.domain.User;
 import com.webtoon.core.user.domain.enums.Gender;
-import com.webtoon.core.user.domain.UserRepository;
+import com.webtoon.core.user.repository.UserRepository;
 import com.webtoon.core.episode.domain.Episode;
 import com.webtoon.core.webtoon.domain.Webtoon;
 import com.webtoon.core.webtoon.domain.enums.EndFlag;
 import com.webtoon.core.webtoon.domain.enums.StoryGenre;
 import com.webtoon.core.webtoon.domain.enums.StoryType;
-import com.webtoon.core.episode.domain.EpisodeRepository;
+import com.webtoon.core.episode.repository.EpisodeRepository;
 import com.webtoon.core.comment.domain.Comment;
-import com.webtoon.core.comment.domain.CommentDislikeRepository;
-import com.webtoon.core.comment.domain.CommentLikeRepository;
-import com.webtoon.core.comment.domain.CommentRepository;
+import com.webtoon.core.comment.repository.CommentDislikeRepository;
+import com.webtoon.core.comment.repository.CommentLikeRepository;
+import com.webtoon.core.comment.repository.CommentRepository;
 import com.webtoon.core.comment.dto.CommentsResponse;
 import com.webtoon.core.comment.dto.MyPageCommentResponse;
 import com.webtoon.core.comment.dto.MyPageCommentsResponse;
